Migrate samples to Apollo v2 and re-generate GraphQL code
Upgrade sample apps to Apollo v2 and re-run code generation. Apollo v1 generated code caused Swift 6 concurrency issues, so the generated GraphQL fragments and operations were regenerated with the v2 generator (formatting/structure changes only). The Xcode project references were updated to remove obsolete generated entries and align with the new code layout. This change removes the temporary workaround and enables the 2025-07 Swift upgrade / follow-up PR to proceed.

+++ Samples/MobileBuyIntegration/MobileBuyIntegration/Views/ProductView.swift
@@ -238,7 +238,7 @@ class ProductCache: ObservableObject {
             let network = Network.shared
 
             let query = Storefront.GetProductsQuery(
-                first: .some(limit),
+                first: .some(Int32(limit)),
                 country: network.countryCode,
                 language: network.languageCode
             )
